Update README.md
adds a README file detailing the background, objectives, project scope, team members and roles, as well as the functionalities and characteristics of the DisApp24 reservation module. This will provide a clear understanding of the project for stakeholders and collaborators. The README also outlines the task distribution within the project team and how the resource reservation module will expand the functionalities of DisApp24. This README will be updated as the project progresses.