Add RSS and Calendar tab icons to Resources/AppIcon
This commit adds the RSS and Calendar tab icons to the Resources/AppIcon folder. These icons will be used in the AppShell.xaml file to represent the respective tabs in the app's user interface.

The addition of these icons improves the user experience by providing clear visual cues for navigating between the RSS and Calendar tabs.