Display full content of selected Rss item in detailed content page
Update OnItemSelected method in RssFeedView to display the full content of the selected Rss item in the detailed content page. This fix addresses issue #1(https://github.com/mardausdennis/ReservationmoduleDisApp24/issues/1), which prevented users from fully engaging with the content and defeated the purpose of having a detailed content page. The method now retrieves the content from the "encoded" element and removes any HTML tags before displaying it in a new StackLayout. This change improves the usability and functionality of the RssFeedView.