Fix issue with long content in ContentPage not being scrollable
This commit addresses issue #2, which reported that when the content of the ContentPage is too long, it is not possible to scroll down and see the entire content. To fix this, a ScrollView has been added to the ContentPage, which allows the content to be scrollable when it exceeds the size of the screen.