Fix issue with XCalendar package not allowing selection of individual dates
Previously, the XCalendar package used in our Maui project was not allowing users to select individual dates, which was causing issues for our calendar display. After investigating the problem, I found that there was a bug in the code that was preventing the selection of individual dates. I fixed this bug by modifying the code to allow for individual date selection. This change should improve the user experience for our calendar display and make it more functional.