Remove unnecessary Models folder
Previously, our project included a Models folder that was no longer being used and had become unnecessary. This folder was taking up space and cluttering the project directory, so I decided to remove it. The removal of the Models folder will not affect the functionality of the project in any way. T